backend/pyproject.toml
[project]
name = "vectorizer-backend"
version = "0.1.0"
description = "Backend para vectorizador de imágenes RGB a SVG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "opencv-python-headless>=4.9.0",
    "pillow>=10.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel.packages]
path = "src"